Separated processing from stitching, MP bug fix
Moved preprocessing from generate_stitch.py to preprocess.py, updated run_forecast.sh to match.  Changed imap back to map in generate_stitch.py because imap wasn't executing under some circumstances.  (Args passed to map should be only metadata anyway, so memory impact should be minimal.)  All MP functions should be further tested/optimized in the future for both speed and memory usage.
