+++ code/scripted_processing/config.conf
@@ -61,7 +61,7 @@ GHI_Coor = {1:   [40.868972, -72.852225],
 #days=['20191013','20191014','20191015','20191016']
 #days=['20191014','20191015','20191016']
 
-days=['20200101']
+days=['20190101','20190102','20190103','20190104','20190105','20190106','20190107']
 lead_minutes=[1,3,5,10,15,30,45]
 sensors=[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25]
 forecast_timezone=EST
